authenticate.ts , token.ts 수정
