+++ package.json
@@ -17,6 +17,7 @@
     "@types/cookie-parser": "^1.4.10",
     "@types/cors": "^2.8.19",
     "@types/express": "^5.0.6",
+    "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^25.0.1",
     "nodemon": "^3.0.1",
     "typescript": "^5.9.3"
