+++ src/types/express.d.ts
@@ -1,0 +1,11 @@
+import type { User } from '@prisma/client';
+
+declare global {
+  namespace Express {
+    interface Request {
+      user?: User;
+    }
+  }
+}
+
+export {};
